requirements.txt
streamlit==1.21.0
pandas==2.0.1
gspread==5.7.1
google-auth==2.23.0
google-auth-oauthlib==0.7.1
google-auth-httplib2==0.1.0
aiohttp==3.8.4
requests==2.28.2
